capture: share the tun fixture
The dispatcher's tests need a raw IP link too.
